move memory_patcher to qt_gui
