+++ src/emulator.cpp
@@ -8,7 +8,7 @@
 #include "common/logging/backend.h"
 #include "common/logging/log.h"
 #ifdef ENABLE_QT_GUI
-#include "common/memory_patcher.h"
+#include "qt_gui/memory_patcher.h"
 #endif
 #include "common/ntapi.h"
 #include "common/path_util.h"
