+++ src/qt_gui/cheats_patches.cpp
@@ -28,7 +28,7 @@
 #include <QXmlStreamReader>
 #include <common/logging/log.h>
 #include "cheats_patches.h"
-#include "common/memory_patcher.h"
+#include "qt_gui/memory_patcher.h"
 #include "common/path_util.h"
 #include "core/module.h"
