+++ src/qt_gui/memory_patcher.h
